Fallback to magma in tropicalVarietyNewton
Put main part of tropicalVarietyNewton into tropicalVarietyNewtonTriang
to simplify the fallback to magma.

Make expectedValuation return nothing instead of an error, pass this
back to tropicalVarietyNewtonTriang if necessary and call
tropicalVarietyNewtonWithMagma with the current triangular set in this
case.

This removes the need to call tropicalVarietyNewtonWithMagma directly in
tropicalLinkNewton, thus (hopefully) reducing the calls to magma.